feat: move Jackson3Serializer to core module
Jackson 3 support previously lived only in the Spring Boot 4 starter,
which made it unusable outside Spring. Move it into the core serializer
package alongside JacksonSerializer (Jackson 2), with Jackson 2 and 3 as
side-by-side optional dependencies.

Fixes #777
