+++ db-scheduler/src/main/java/com/github/kagkarlsson/scheduler/serializer/jackson3/InstantDeserializer.java
@@ -11,7 +11,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.kagkarlsson.scheduler.boot.autoconfigure.jackson3;
+package com.github.kagkarlsson.scheduler.serializer.jackson3;
 
 import java.time.Instant;
 import tools.jackson.core.JacksonException;
