+++ db-scheduler/src/test/java/com/github/kagkarlsson/scheduler/serializer/jackson3/Jackson3SerializerTest.java
@@ -1,7 +1,22 @@
-package com.github.kagkarlsson.scheduler.boot.autoconfigure;
+/*
+ * Copyright (C) Gustav Karlsson
+ *
+ * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
+ *
+ * <p>http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * <p>Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.kagkarlsson.scheduler.serializer.jackson3;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import com.github.kagkarlsson.scheduler.serializer.Jackson3Serializer;
+import com.github.kagkarlsson.scheduler.serializer.jackson.ScheduleAndDataForTest;
 import com.github.kagkarlsson.scheduler.task.ExecutionComplete;
 import com.github.kagkarlsson.scheduler.task.helper.PlainScheduleAndData;
 import com.github.kagkarlsson.scheduler.task.schedule.CronSchedule;
@@ -44,7 +59,6 @@ public class Jackson3SerializerTest {
     Instant now = Instant.now();
     Daily daily = Schedules.daily(LocalTime.MIDNIGHT);
     byte[] serialize = serializer.serialize(daily);
-    System.out.println(new String(serialize));
     Daily deserialized = serializer.deserialize(Daily.class, serialize);
     assertEquals(daily.getInitialExecutionTime(now), deserialized.getInitialExecutionTime(now));
   }
@@ -54,7 +68,6 @@ public class Jackson3SerializerTest {
     Instant now = Instant.now();
     FixedDelay fixedDelay = Schedules.fixedDelay(Duration.ofHours(1));
     byte[] serialize = serializer.serialize(fixedDelay);
-    System.out.println(new String(serialize));
     FixedDelay deserialized = serializer.deserialize(FixedDelay.class, serialize);
     assertEquals(
         fixedDelay.getNextExecutionTime(ExecutionComplete.simulatedSuccess(now)),
