+++ pom.xml
@@ -73,6 +73,7 @@
     <micrometer.version>1.16.5</micrometer.version>
     <gson.version>2.14.0</gson.version>
     <jackson.version>2.21.3</jackson.version>
+    <jackson3.version>3.1.2</jackson3.version>
     <postgresql.version>42.7.11</postgresql.version>
     <guava.version>33.6.0-jre</guava.version>
   </properties>
